tclExecute: Fix for CHERI by not violating object allocation bounds
This is needed to support CHERI, and thus Arm's experimental Morello
prototype, where pointers are implemented using unforgeable capabilities
that include bounds and permissions metadata to provide fine-grained
spatial and referential memory safety, as well as revocation by sweeping
memory to provide heap temporal memory safety.

In the case that GrowEvaluationStack allocates a new stack, subtracting
TD is strictly UB, since pointer subtraction is only defined within an
object, not between objects. However, the bigger issue is that this is
then added to TD-derived pointers to get equivalent pointers into the
new stack, as well as reconstruct the new TD, but this is UB as pointer
addition is only defined when the result is within the same object (or
one-past-the-end). CHERI detects this and faults trying to update
catchTop since the (hidden) dereference of TD is outside the bounds of
the original TD's allocation.

Fix this by deriving pointers from the new allocation, not the old one.